Documentation page scope & tabs over TOC
Removed the Documentation page in favor of splitting its core sections
into separate pages: Global configuration, Methods, and Local
attributes; This removes the need for a table of contents, as each page
will not nearly be as long. In addition, tabs will be used on the
documentation pages to further reduce the amount of scrolling while
keeping the page tidy.