Configuration page updates
Decided to include code snippets for each tab describing the different
keys/values that can be used within the global configuration object;
This is to narrow the scope of the tab while providing useful code
snippets and examples for when a key can have different types of values.

In addition, some additional documentation was added, particularly for
keys/values that had an ambiguous explaination and required a thorough
breakdown for what each of its values corresponds to.