Code component mousetip integration
Changed the Code component to not use mousetip by default; This is
because the only Code component with a live mousetip example is on the
start page, whereas every other instance is just to show JS code.

Updated the "Getting started" page's "What's a mousetip?" tab to account
for this change and explitly set `mousetip={true}` on the Code component
used.