Updates
Made updates to better align this documentation site with that of
chassis.joeleisner.com, what this site is based off of. This includes
providing an `.nvmrc`, updating dependencies, grouping component styles
and javascript together, switching to SCSS and DartSass instead of SASS
and NodeSass, and more.

In addition, some pages were altered to reflect the recent CSS module
change to the 3.0.0 branch of `purejs-mousetip` and the repo now pulls
in the alpha release of `purejs-mousetip`.