Production build fixes
It came to my attention that the `build` NPM script was failing;
Specifically, it would fail on the `build-html` stage due to
`purejs-mousetip` expecting to be run within the context of a browser.
This was fixed by reintroducing the `gatsby-node.js` file and adding a
WebPack configuration for that stage to not use any loaders when it
comes across the `purejs-mousetip` package.