Removing mousetip styles
It was noticed that when navigating to and from the homepage, where a
mousetip example with style overrides exists, the style overrides were
not being removed before leaving the page.

An attempt was made to fix this by stopping mousetip functionality when
the card or card component are unmounting, but this did not work. As a
work around, the global styles and style overrides are removed in
addition to mousetip functionality being stopped when unmounting.